Move JUnitXMLReporting to codenameone-javase module
